% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustRCoreFunction.R
\name{clustr}
\alias{clustr}
\title{ClustR}
\usage{
clustr(id = NULL, status = NULL, date = NULL, lat = NULL, long = NULL,
  time_incs = c(90, 180, 365), space_incs = c(0.025, 0.05, 0.1, 0.5),
  n = NULL, bootstraps = 5, data)
}
\arguments{
\item{id}{character: The name of the variable containing the unique identification number for subjects.}

\item{status}{character: The name of the variable containing the disease status of subjects, where 1 indicates cases and 0 indicates controls.}

\item{date}{character: The the name of the variable containing the date associated with disease status, such as diagnosis date or birth date. Date must be in either Y-m-d or Y/m/d format.}

\item{lat}{character: The name of the variable (as a character string) containing the latitude associated with each subject, such as diagnosis address or birth address.}

\item{long}{character: The name of the variable (as a character string) containing the longitude associated with each subject, such diagnosis address or birth address.}

\item{time_incs}{(Optional) numeric: A single number or vector of numbers representing the maximum difference in time (in days) between subjects in clusters. Defaults to c(90,180,365). If a vector of numbers is given, ClustR will try each constraint and return the most significant cluster among them.}

\item{space_incs}{(Optional) numeric: A single number or vector of numbers representing the maximum difference in space (in decimal degrees) between subjects in clusters. Defaults to c(.025, .05, .1, .5). If a vector of numbers is given, ClustR will try each constraint and return the most significant cluster among them.}

\item{n}{(Optional) numeric: A single number representing the size of samples you want ClustR to bootstrap. Defaults to nrow(data) if nrow(data) < 1001 or else defaults to 1000.}

\item{bootstraps}{(Optional) numeric: A single number representing how many samples ClustR should bootstrap. Defaults to 5.}

\item{data}{dataframe: The name of the dataset ClustR should evaluate.}
}
\description{
This function evaluates space-time clustering of case-control or cohort data. ClustR simulates a random distribution of disease for the input data and then compares the true data against it by bootstrapping samples and comparing chi-square distributions. This function returns a dataframe summarizing which areas and timeframes of the true data were sampled and which yielded unusually high proportions of cases or controls compared to the simulation. It also returns p-values for each sample (generated_p) and then adjusts for false discovery rate via the Benjamini-Hochberg approach (FDR_p).
}
\examples{
\dontrun{
clustr_test1 <- clustr(id="blindid",status="case_status",date="birthdate",lat="mom_lat",
       long="mom_long",time_incs=c(30,90),space_incs=c(.5,1),n=500,bootstraps=20,data=testdata)
}
}
\keyword{clustering}
\keyword{space}
\keyword{time}
