% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClustRFunction.R
\name{get.clustr}
\alias{get.clustr}
\title{Get ClustR}
\usage{
get.clustr(data = NULL, id = NULL, status = NULL, date = NULL,
  lat = NULL, long = NULL, results = NULL, type = 1, adj = FALSE,
  pctl.cutoff = 1, adj.cutoff = 0.05, space_incs = NULL)
}
\arguments{
\item{data}{dataframe: The name of the original dataset clustr() used to produce results.}

\item{id}{character: The name of the variable in data containing the unique identification number for subjects.}

\item{status}{character: The name of the variable in data containing the disease status of subjects, where 1 indicates cases and 0 indicates controls.}

\item{date}{character: The the name of the variable in data containing the date associated with disease status, such as diagnosis date or birth date. Date must be in either Y-m-d or Y/m/d format.}

\item{lat}{character: The name of the variable in data (as a character string) containing the latitude associated with each subject, such as diagnosis address or birth address.}

\item{long}{character: The name of the variable in data (as a character string) containing the longitude associated with each subject, such diagnosis address or birth address.}

\item{results}{dataframe: The name of the clustr() output.}

\item{type}{binary: Either 0 or 1 to indicate whether you want to map control clusters or case clusters, respectively. Defaults to 1.}

\item{adj}{logical: Indicates whether you want data from clusters based on chi-square values (unadjusted) or based on the Benjamini-Hochberg p-values (adjusted). Defaults to FALSE.}

\item{pctl.cutoff}{categorical (1,2,3): If you are not mapping clusters based on the adjusted p-value, you must specify what percentile cut-off to use. 1 indicates whether you want clusters with a chi-square value above the 100th percentile, 2 for above the 99th percentile, 3 for above the 97.5th percentile. Defaults to 1.}

\item{adj.cutoff}{numeric: If using post-adjustment clusters, indicate what the significance cut-off should be. Defaults to 0.05.}

\item{space_incs}{(Optional) numeric: A single number or vector of numbers representing what size (in decimal degrees) of clusters you want to pull subjects from.}
}
\description{
This function generates dynamically named CSVs with the IDs and disease statuses of subjects within disease clusters identified by clustr().
}
