% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapClustRFunction.R
\name{map.clustr}
\alias{map.clustr}
\title{Map ClustR Results}
\usage{
map.clustr(results = NULL, type = 1, adj = FALSE, adj.cutoff = 0.05,
  zoom = 6, region = NULL)
}
\arguments{
\item{results}{dataframe: The name of the clustr() output to map from.}

\item{type}{binary: Either 0 or 1 to indicate whether you want to map control clusters or case clusters, respectively. Defaults to 1.}

\item{adj}{logical: Indicates whether you want to plot clusters based on chi-square values (unadjusted) or based on the Benjamini-Hochberg p-values (adjusted). Defaults to FALSE.}

\item{adj.cutoff}{numeric: If mapping post-adjustment clusters, indicate what the significance cut-off should be. Defaults to 0.05.}

\item{zoom}{numeric: Tells ggmap how far to zoom into your region. Defaults to 6.}

\item{region}{character: Name of the region (e.g. state, city, country) where your clusters are.}
}
\description{
This function maps the radii containing clusters identified by the clustr() function using ggmap.
}
\examples{
\dontrun{
map.clustr(results=clustr_test5, type=1, adj=TRUE, adj.cutoff=.10, zoom=6, region='California')
}
}
\keyword{clustering}
\keyword{clustr}
\keyword{ggmap}
\keyword{ggplot2}
\keyword{space}
\keyword{time}
